requirements.txt
djangorestframework>=3.15.2,<3.16
psycopg[binary, pool]>=3.2.1,<3.3
channels>=4.1.0,<4.2
channels-redis>=4.2.0,<4.3
daphne>=4.1.2,<4.2
drf-spectacular>=0.27.2,<0.28
djangorestframework-simplejwt>=5.3.1,<5.4
pytest-asyncio>=0.24.0,<0.25
pytest-django>=4.9.0,<4.10
django-cors-headers>=4.5.0,<4.6